+++ src/components/index.ts
@@ -55,3 +55,4 @@ export * from './topVehicleSharedBadge';
 export * from './discreteSlider';
 export * from './formControlSection';
 export * from './rangeSlider';
+export { SearchableList, ListItemWithChildren } from './list/SearchableList';
