+++ src/components/list/SearchableListItem.tsx
@@ -1,12 +1,14 @@
 import React, { FC, PropsWithChildren } from 'react';
 import { RangeTuple } from 'fuse.js';
-import { Button, CheckboxCheckedChangeDetails } from '@chakra-ui/react';
+import {
+  Button,
+  ButtonProps,
+  CheckboxCheckedChangeDetails,
+  List,
+} from '@chakra-ui/react';
 
 import { Checkbox, CheckboxProps } from '../checkbox';
-
 import { SearchableListItemLabel } from './SearchableListItemLabel';
-
-import { List } from './index';
 
 type CommonListItem = {
   label: string;
@@ -16,11 +18,12 @@ type CommonListItem = {
   showChevron?: boolean;
   highlightIndices?: readonly RangeTuple[];
   isCheckbox?: boolean;
+  paddingLeft?: string;
 };
 
 type CommonProps = {
   value: string;
-  paddingY: CheckboxProps['paddingY'];
+  paddingY: ButtonProps['paddingY'] | CheckboxProps['paddingY'];
   name: string;
   'aria-current': boolean;
 };
@@ -32,7 +35,7 @@ type CheckboxListItem = {
 } & CommonListItem;
 
 type RadioButtonListItem = {
-  onClick: React.MouseEventHandler<HTMLButtonElement>;
+  onClick: ButtonProps['onClick'];
   isCheckbox?: false;
 } & CommonListItem;
 
@@ -82,26 +85,21 @@ export const SearchableListItem: FC<PropsWithChildren<ListItemType>> = (
     variant: 'alignTop',
     onChange: isCheckbox ? props.onClick : undefined,
   };
+  const radioButtonProps: ButtonProps = {
+    ...commonProps,
+    onClick: !isCheckbox ? props.onClick : undefined,
+    onChange: undefined,
+    width: 'full',
+    display: 'flex',
+    paddingX: 'none',
+  };
 
   return (
-    <List.Item css={{ breakInside: 'avoid' }}>
+    <List.Item css={{ breakInside: 'avoid' }} paddingLeft={props.paddingLeft}>
       {isCheckbox ? (
         <Checkbox {...checkboxProps} />
       ) : (
-        <Button
-          value={value}
-          paddingY="sm"
-          name={`searchable-list-item-${value}`}
-          aria-current={isSelected}
-          onClick={!isCheckbox ? props.onClick : undefined}
-          width="full"
-          display="flex"
-          paddingX="0"
-          css={{
-            background: 'transparent',
-            _hover: { background: 'gray.100' },
-          }}
-        >
+        <Button {...radioButtonProps}>
           <SearchableListItemLabel {...labelProps} />
         </Button>
       )}
